+++ user/userController.py
@@ -39,28 +39,9 @@ def login():
     else :
         print("login fail")
         return render_template('user.html')
+#=======================================================================================================
 
-# 글쓰기 페이지 (GET)
-@user.route('/thread/write', methods=['GET'])
-def writePage():
-    return render_template('thread_write.html')
 
-# 글 작성 처리 (POST)
-@user.route('/thread/write', methods=['POST'])
-def writePost():
-    db = get_db("sk27")           # DB 이름: sk27
-    threads = db['thread']        # thread 컬렉션
-
-    title = request.form['title']
-    body = request.form['body']
-
-    threads.insert_one({
-        "threadTitle": title,
-        "threadBody": body,
-        "uploadDate": datetime.utcnow()
-    })
-
-    return redirect(url_for('user.allThreads'))
 
 # 글 전체 보기 (목록)
 @user.route('/thread/all', methods=['GET'])
